% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotSet.R
\name{boxplot_combine}
\alias{boxplot_combine}
\title{Combined Box plots for all data sets}
\usage{
boxplot_combine(gene_info, title, ana, fontsize_plot, fontsize_pvalue,
  fontsize_legend)
}
\arguments{
\item{gene_info}{A character or a number. The gene information. Can be either gene name or gene ID. Gene name can be uppercase or lowercase. For example: The gene name can be "S100A8" or "s100a8".}

\item{title}{A character. The description that should be put on the combined boxplot. The example uses GENE name.}

\item{ana}{TRUE or FALSE. Whether We should put p-value information on the plot. If ana = T then the p-vlaue info will be placed on the boxplot, if ana=FALSE, the p-value info will not be placed on the boxplots.}

\item{fontsize_plot}{A number. The font size of plot(title, caption, legend title ... and so on)}

\item{fontsize_pvalue}{A number. The font size of p-value (comparison of two groups and Omnibus p-value) on the plot.}
}
\description{
Combine box plots in five different data sets.
}
\details{
We will use function to put boxplots together for five different data sets.
}
\examples{
\dontrun{
boxplot_combine(29126,"CD274",ana=TRUE,fontsize_plot = 10, fontsize_pvalue = 5, fontsize_legend = 10)
boxplot_combine("CD274","CD274",ana=FALSE,fontsize_plot = 10, fontsize_pvalue = 5, fontsize_legend = 10)
}


}
