% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{cutoff}
\title{cutoff}
\usage{
cutoff(datasurv, gene_expr, criterion)
}
\arguments{
\item{datasurv}{A dataset. A dataset contains time and event information.}

\item{gene_expr}{A vector. A vector contains gene expression information.}

\item{criterion}{A number. A criterion that is used to split two groups. Range from 0 to 1.}
}
\description{
Group the data based on log rank test.
}
\details{
The data set is grouped by the p-value of log-rank test. \cr
First: We will order the gene expression.\cr
Second: Divide them as two group for any possibility, the minimal size of each group is 10 percent of total sample size.\cr
Third: Based on the group, we will have the p-values of log-rank test for each possibility \cr
Fourth: Order the p-values from minimal to maximum. \cr
Fifth: The smallest p-value will be the cutoff point. The data will be separated based on that cutoff point. If gene expression is bigger than or equal to the gene expression of cutoff point, then groups 1, otherwise groups 0. \cr
Then it will create a KM plot based on the group.
}
