% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_translate.R
\name{gene_translate}
\alias{gene_translate}
\title{Gene tranlator}
\usage{
gene_translate(gene_info)
}
\arguments{
\item{gene_info}{A character or a number. The gene information. Can be either gene name or gene ID. Gene name can be uppercase or lowercase. For example: The gene name can be "S100A8" or "s100a8".}
}
\description{
It can automatically translate gene names to gene ID.
}
\details{
When you type in no matter gene ID or gene name, this function will automatically give you gene ID eventually.
}
\examples{
\dontrun{
gene_translate(6279)
gene_translate("s100A8")
gene_translate("S100A8")

}

}
