% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMplot.R
\name{kmplot}
\alias{kmplot}
\title{K-M plot for genes}
\usage{
kmplot(dataset, gene_info, time, event, receptor, criterion, title,
  fontsize)
}
\arguments{
\item{dataset}{A data set. The data set used.}

\item{gene_info}{A character or a number. The gene information. Can be either gene name or gene ID. Gene name can be uppercase or lowercase. For example: The gene name can be "S100A8" or "s100a8".}

\item{time}{A character. The survival time variable name in original dataset.}

\item{event}{A character. The event variable name in original dataset.}

\item{receptor}{A logical indicating whether we will use the whole data set or receptor status = TN only. If receptor = TRUE, then use the data with receptor status = TN only; if receptor = FLASE, then use the whole data set.}

\item{criterion}{A number. A criterion that is used to split two groups. Range from 0 to 1.}

\item{title}{A character. The title of the KM plot.}

\item{fontsize}{A number. The font size. The default is 1.}
}
\description{
K-M plots for a specified gene and a specified data set.
}
\details{
In this function, it will create a dataset with three variables.
The first column of data set is time variable, the second column is censored or not, the third column is group variable.
The data set is grouped by the p-value of log-rank test. \cr
First: We will order the gene expression.\cr
Second: Divide them as two group for any possibility, the minimal size of each group is 10 percent of total sample size.\cr
Third: Based on the group, we will have the p-values of log-rank test for each possibility \cr
Fourth: Order the p-values from minimal to maximum. \cr
Fifth: The smallest p-value will be the cutoff point. The data will be separated based on that cutoff point. If gene expression is bigger than or equal to the gene expression of cutoff point, then groups 1, otherwise groups 0. \cr
Then it will create a KM plot based on the group.
}
\examples{
\dontrun{
kmplot(ispy1, gene_info = 6279, "rfs.t", "rfs.e",receptor=T,criterion=0.2, title="ISPY1",fontsize=20)
kmplot(ispy1, gene_info = "S100A8", "rfs.t", "rfs.e",receptor=T, title="ISPY1")
kmplot(ispy1, gene_info = "s100a8", "rfs.t", "rfs.e",receptor=T, title="ISPY1")

}

}
