% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneBoxPlot.R
\name{oneboxplot}
\alias{oneboxplot}
\title{Box plots by receptor types}
\usage{
oneboxplot(dataset, gene_info, title, trans, ana, fontsize_plot,
  fontsize_pvalue, fontsize_legend)
}
\arguments{
\item{dataset}{A data set. The data set used in the boxplots.}

\item{gene_info}{A character or a number. The gene information. Can be either gene name or gene ID. Gene name can be uppercase or lowercase. For example: The gene name can be "S100A8" or "s100a8".}

\item{title}{A character. The boxplot's title, can be specified as gene's name}

\item{trans}{TRUE or FALSE. Do log transformation or not. If trans = TRUE, then the gene expression will be tranformed to the base 2 logarithm. If trans = FLASE, then it will use original gene expression values.}

\item{ana}{TRUE or FALSE. Whether We should put p-value information on the plot. If ana = T then the p-vlaue info will be placed on the boxplot, if ana=FALSE, the p-value info will not be placed on the boxplots.}
}
\description{
Comparing the gene expresstions based on receptor types by showing each boxplot together.
}
\details{
This function will create boxplots based on receptor type (HER2, ERPR, TN) for a gene and a data set. Besides boxplots it also
gives us the t-test p-value of possible receptor type combinations and the ANOVA p-value of all receptor types. The p-values,
total N, and the means will automatically show on the plot.
}
\examples{
\dontrun{
oneboxplot(yau,5293,"YAU",trans = F, ana = T, fontsize_plot = 10, fontsize_pvalue = 5, fontsize_legend = 10)
oneboxplot(yau,5293,"YAU",trans = F, ana = F, fontsize_plot = 10, fontsize_pvalue = 5, fontsize_legend = 10)
}
}
\seealso{
\link{cor_plot}
}
